move parse_partition_table from cli to flasher mod
